requirements.txt
pydantic>=2.7.0
openai>=1.40.0
openenv>=0.1.0
PyYAML>=6.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
